code/Amazon_AI_Challenge/ai-agent/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.1
av==15.0.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
cyclopts==3.22.2
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
docutils==0.21.2
email_validator==2.2.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.115.13
fastmcp==2.10.6
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.154.0
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.27.0
google-genai==1.26.0
googleapis-common-protos==1.70.0
h11==0.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
livekit==1.0.12

# MongoDB & Database
pymongo==4.9.2
motor==3.6.0

# Vector Database & RAG
# chromadb==0.5.23
pypdf==5.1.0
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-community==0.3.14

# Google Cloud additional
google-cloud-vision==3.8.1

# Google Sheets (kept for optional usage in other flows)
gspread==6.1.4
google-auth==2.40.3
google-auth-oauthlib==1.2.0

# Utilities
python-dotenv==1.1.1
livekit-agents==1.2.2
livekit-api==1.0.4
livekit-blingfire==1.0.0
livekit-plugins-google==1.2.1
livekit-plugins-simli==1.2.2
livekit-protocol==1.0.4
markdown-it-py==3.0.0
mcp==1.12.0
mdurl==0.1.2
multidict==6.6.3
nest-asyncio==1.6.0
numpy==2.3.1
openai==1.90.0
openapi-pydantic==0.5.1
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
pillow==11.3.0
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.9.0
python-decouple==3.8
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
referencing==0.36.2
reportlab==4.4.2
requests==2.32.4
rich==14.0.0
rich-rst==1.3.1
rpds-py==0.26.0
rsa==4.9.1
sniffio==1.3.1
sounddevice==0.5.2
sse-starlette==2.4.1
starlette==0.46.2
tenacity==8.5.0
tqdm==4.67.1
twilio==9.6.5
types-protobuf==4.25.0.20240417
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
